requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.88.0
h11==0.14.0
idna==3.4
picamera==1.13
pydantic==1.10.2
python-dotenv==0.21.0
RPi.GPIO==0.7.1
sniffio==1.3.0
starlette==0.22.0
typing-extensions==4.4.0
uvicorn==0.20.0
websockets==10.4
